+++ examples/base_1cc/variables.tf
@@ -225,3 +225,13 @@ variable "zssupport_server" {
   description = "destination IP address of Zscaler Support access server. IP resolution of remotesupport.<zscaler_customer_cloud>.net"
   default     = "199.168.148.101" #for commercial clouds
 }
+
+variable "fips_enabled" {
+  type        = string
+  description = "Enable FIPS mode for Cloud Connector provisioning. Supported values are 'False' or 'True'."
+  default     = "False"
+  validation {
+    condition     = var.fips_enabled == "False" || var.fips_enabled == "True"
+    error_message = "Variable fips_enabled must be either 'False' or 'True'."
+  }
+}